Fix axis rotation.
Blender uses a rest mode with rotations set to 0 degrees so in order to do be able
to do skinnig with Blender we need the skeleton to be set in a proper initial
position with 0 rotations. That leads to bones not being align with an axis.
To fix that on the rendering we show artificial rotation axis intuitive for the user
eg along the bone but internally we maintain arbitrary rotations.