Added class PointCloud
In an effort to have a more robust code and use PCL the cloud points in
osg::Vec3Array have been substituted by the PointCloud class. With this class
the cloud implementation is hidden and the points can be accessed with any of
the vector types used in the program.